// packages/datadog-api-client-v1/models/WidgetFormula.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WidgetFormulaLimit } from './WidgetFormulaLimit';
/**
* Formula to be used in a widget query.
*/
export declare class WidgetFormula {
    /**
    * Expression alias.
    */
    'alias'?: string;
    /**
    * String expression built from queries, formulas, and functions.
    */
    'formula': string;
    'limit'?: WidgetFormulaLimit;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): WidgetFormula;
    static serialize(data: WidgetFormula): {
        [key: string]: any;
    };
    constructor();
}