// packages/datadog-api-client-v1/models/WidgetRequestStyle.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WidgetLineType } from './WidgetLineType';
import { WidgetLineWidth } from './WidgetLineWidth';
/**
* Define request widget style.
*/
export declare class WidgetRequestStyle {
    'lineType'?: WidgetLineType;
    'lineWidth'?: WidgetLineWidth;
    /**
    * Color palette to apply to the widget.
    */
    'palette'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): WidgetRequestStyle;
    static serialize(data: WidgetRequestStyle): {
        [key: string]: any;
    };
    constructor();
}